refactor(editor): clean up code and add annotations for DTOs
- Removed unnecessary blank line in `getResult` method.
- Added `@Builder` annotation to `Status`.
- Fixed formatting issue in `LanguageResponse`.
- Changed `time` field type in `SubmissionResult` from `Double` to `Float`.

+++ src/main/java/com/icoder/coding/editor/dto/SubmissionResult.java
@@ -15,7 +15,7 @@ import lombok.NoArgsConstructor;
 public class SubmissionResult {
     private String stdout;
 
-    private Double time;
+    private Float time;
 
     private Integer memory;
